refactor components
