--- app/src/features/rules/screens/rulesList/components/RulesList/components/CharlesImporter/charlesImporter.css
@@ -1,168 +1,0 @@
-.import-from-charles-modal .ant-modal-body .rq-modal-content:last-child {
-  border-bottom-left-radius: inherit;
-  border-bottom-right-radius: inherit;
-}
-
-.charles-import-wrapper {
-  width: 55%;
-  margin: 0 auto;
-  max-width: 900px;
-  background-color: transparent;
-  height: 100%;
-  padding: 24px;
-}
-
-.charles-import-container {
-  margin: auto;
-}
-
-.charles-import-container .charles-import-actions-row {
-  display: flex;
-  justify-content: flex-end;
-  margin: 16px 0;
-  gap: 8px;
-}
-
-.charles-import-header {
-  margin-bottom: 8px;
-  margin-top: -18px;
-}
-
-.charles-import-body {
-  background: var(--requestly-color-surface-1);
-  margin-bottom: 16px;
-  border-radius: 8px;
-  padding: 16px;
-}
-
-.import-from-charles-modal .charles-import-footer {
-  color: var(--requestly-color-text-subtle);
-  font-size: var(--requestly-font-size-xs);
-  text-align: center;
-  padding-top: 16px;
-  border-top: 1px solid var(--requestly-color-white-t-20);
-  margin-top: 16px;
-  width: 550px;
-  margin-left: -24px;
-}
-
-.charles-import-footer {
-  color: var(--requestly-color-text-subtle);
-  font-size: var(--requestly-font-size-xs);
-  text-align: center;
-  margin: 16px 0;
-}
-
-.charles-import-heading {
-  font-size: var(--requestly-font-size-base);
-  font-weight: var(--requestly-font-weight-medium);
-  gap: 8px;
-  display: flex;
-  align-items: center;
-}
-
-.charles-import-heading .charles-import-back-icon {
-  cursor: pointer;
-}
-
-.charles-import-share-container {
-  font-size: var(--requestly-font-size-xs);
-  font-weight: var(--requestly-font-weight-medium);
-  padding: 8px;
-  cursor: pointer;
-  margin-top: 4px;
-  margin-right: 20px;
-}
-
-.parsed-rules-info {
-  display: flex;
-  /* justify-conxtent: center; */
-  flex-direction: column;
-  /* padding: 16px; */
-}
-
-.doc-link-container {
-  margin-left: 18px;
-}
-
-.charles-import-warning-banner {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background-color: var(--requestly-color-warning-600);
-  border: 1px var(--requestly-color-white-t-20) solid;
-  border-radius: 4px;
-  padding: 8px 16px;
-  margin-bottom: 16px;
-  color: var(--requestly-color-black);
-}
-
-.charles-import-warning-banner-text {
-  font-size: var(--requestly-font-size-sm);
-  gap: 8px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.charles-importer-warning-icon {
-  margin-top: -2px;
-}
-
-.charles-importer-warning-docs-link a {
-  color: var(--requestly-color-black);
-  text-decoration: underline;
-}
-
-.parsed-success-row {
-  color: var(--requestly-color-success-300);
-  gap: 8px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.parsed-rules-info li {
-  list-style: disc;
-}
-
-.external-icon-link {
-  vertical-align: text-top;
-}
-
-.info-link-container {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  color: var(--requestly-color-error-300);
-}
-
-.parsed-rules-error-info {
-  font-size: var(--requestly-font-size-base);
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  padding: 0 16px;
-}
-
-.parsed-rules-error-info ol {
-  font-weight: var(--requestly-font-weight-normal);
-  font-size: var(--requestly-font-size-sm);
-}
-
-.parsed-rules-error-info li {
-  list-style: unset;
-}
-
-.validation-heading {
-  color: var(--requestly-color-error-300);
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.validation-subheading {
-  margin-left: 22px;
-  font-weight: var(--requestly-font-weight-medium);
-  margin-bottom: 8px;
-}
