+++ app/src/features/rules/screens/rulesList/components/RulesList/components/ImporterComponents/ModheaderImporter/ImportFromModheaderScreen.tsx
@@ -3,7 +3,7 @@ import { ImportFromModheader } from "./ModheaderImporter";
 
 export const ImportFromModheaderWrapperView: React.FC = () => {
   return (
-    <div className="modheader-import-wrapper">
+    <div className="importer-wrapper">
       <ImportFromModheader />
     </div>
   );
