+++ app/src/features/rules/screens/rulesList/components/RulesList/components/index.ts
@@ -1,6 +1,6 @@
 export { GettingStarted } from "./GettingStarted";
 export { CreateTeamRuleCTA } from "./GettingStarted/CreateTeamRuleCTA";
-export { ImportFromCharlesModal, ImportFromCharlesWrapperView } from "./CharlesImporter";
+export { ImportFromCharlesModal, ImportFromCharlesWrapperView } from "./ImporterComponents/CharlesImporter";
 export { ImportRulesModal } from "../../../../../modals/ImportRulesModal";
 export { RuleTypesDropdownWrapper } from "./RuleTypesDropdownWrapper/RuleTypesDropdownWrapper";
 export { NewRuleButtonWithDropdown } from "./NewRuleButtonWithDropdown/NewRuleButtonWithDropdown";
